MODCAMUNDA-7: Add several delegate unit tests and fix relating problems.
Switch to use `@Spy` instead of `@InjectMocks` in most cases to avoid problems where Mockito has conflicts when abstract-classes or interfaces are used multiple times with `@Mock`.
Add and use `@BeforeEach` in the case where `@InjectMocks` seems to be better served.

The delegate unit tests exposed a design bug where the abstract interfaces are calling a getter without checking if the value is NULL for the default abstract implementation.
There is a "has" function added to child abstract implementation that provides the expected safety check.
Move this child abstract class "has" method into the parent abstract class and update the default abstract getter method to call this "has" method.