MODCAMUNDA-16: Clean up test changes and add back removed exception mock.
The MockMVC constants utility only mocks simply types but multiple UUIDs are needed here.
Add the UUIDs as constants in the test.

Restore the practice of having the variables at the class level.
This better allows for reduction of code when there are multiple tests.

Restore the abstract node, now calling it `simpleNode`.
The unit tests should be isolated.
Using the abstract class for the node allows for a better isolated test.
Keep the start/input/end nodes for now (although these might need to be abstracted out in the future).

The `testFromWorkflowException` is testing the exception.
The exception was previously removed from the test thus nullifying the entire purpose of the test.
Add back the exception, rename the function to `testFromWorkflowExceptionDuringSetupUsingWarnings`, and add some inline comments.
This should help prevent the unwarranted removal of the actual purpose of the unit test.
The mocking of the logger looks complicated and so that has not been done.
Ideally, the log messages should be verified.

Fix white space and spacing/tabbing problems.