MODCAMUNDA-16: Fix the path changes to properly handle the Camunda admin and REST Camunda paths.
The Camunda Admin UI fails to work with the `jersey.application-path` set to `camunda`.
This is due to a misconfiguration.

Change the Camunda path to `ui`.

Now both the OKAPI REST path and the Camunda Admin path exist without conflict.
Both are relative to the context path, which is set to `/`.