Experimentally bring in folio-spring-support-base.
This brings in the `folio-spring-support-base` support needed to get the `/_/tenant` working as expected.
This should help test and confirm that this path is the problem in the Rancher deployment.

The OkHttp code from `mod-workflow` is copied over to get this to work.
Additional `application.yml` configuration settings are also added for testing purposes.