MODCAMUNDA-21: Prepare for spring-boot 3.3.5 support.
Switch to JDK21 which has better spring-boot 3.3.x support as well as better GraalVM support.

The `com.jcraft` has migrated to `com.github.mwiede`, so update that and the version.

Update the graalvm related dependencies.
Add the polyglot dependencies.

The `org.jacoco` `jacoco-maven-plugin` must be upgraded or it will fail with the new spring-boot 3.3.x changes.

Fix some warnings and change the default `springframework.test` setting.