Add SubjectSources and SubjectTypes, updating and cleaning up the unit tests.
The tests are failing because `subjects` is now empty.

This adds the SubjectSources and SubjectTypes "with" and responses.
This does not yet solve the problem but it feels like it is in the correct direction.

The getters are a highly redundant.
Create a private method and use that for the new getters.
Do not change the other functions for now to keep things within scope as much as possible.

The unit test mocks are a massive wall of text.
Simplify the code by using a private method so that I can better read and make sense of this during investigation.
This removes a lot of repitition.