Pull in tenant header name from Spring Module Cores "Spring Tenant".
The previous behavior introduces a regression due to the hardcoded header name.
Bring in the configured settings as per "Spring Tenant".

The `MappingUtility` operates statically outside of Java, such as in Ruby or Javascript.
This puts unusual requirements on the structure of things.
If this were not the case, then the use of non-static Autowired Spring behavior would instead be followed here.

The context is loaded in a non-static manner.
Bring in the Tenant settings content onto a statically stored but modifiable structure.
Then use a static getter to fetch the now updated and statically stored context.
This makes sure to not re-assign the Tenant settings context.

Fix the `OkapiRestTemplate` so that it can be instantiated normally.
Update the unit tests, which now requires Spring.
The context must be manually specified `SpringBootRun` to prevent errors like this:
```
ApplicationContext failure threshold (1) exceeded: skipping repeated attempt to load context for  test profile...
```

The Tenant Header can now be changed dynamically, even when running tests like this:
```shell
TENANT_HEADERNAME="Different-Header" mvn clean test -Dtest=OkapiRestTemplateTest
```