MODCAMUNDA-21: Fix dependencies and suppress warning.
The `-community` edition of the packages should be brought in due to licensing concerns.
Remove the direct dependencies that are already brought in already via the `-community` editions.

Suppress the `graalvm.js` that is brought in from `data-import-processing-core` due to version dependency conflicts.
The version from `graalvm.js` brought in from `data-import-processing-core` is also not the `-community` edition.

Suppress the warning bout dynamically loading bytebuddy as per:
```
WARNING: If a serviceability tool is in use, please run with -XX:+EnableDynamicAgentLoading to hide this warning
```