MODCAMUNDA-22: Update the read me documentation.
Remove a lot of old information or information that would potentially require constant updating.
In cases of constant updating, remove explicit versions or directly link to the source files.

Add build and run documentation.

Remove all of the old vagrant related information.
This ha long since been deprecated and is no longer supported in the FOLIO community, in general.

Fix links as appropriate.