RTR 278-286: Add token utility.
The `HttpCookie` can be directly imported and used to handle cookie-specific tasks.
Therefore, instead of implementing a Cookie Utility, implement a Token Utility to specifically handle the Token-specific aspects.

Provide functions that accept the array of cookies and a function for a single cookie.
Parse the cookie(s) into an `HttpCookie` object.
Expose all of these functions so that the caller can use whichever they need.

The `X-Okapi-Token` is being hard-coded and must not be.
Replace the hard-coded `OKAPI_TOKEN_HEADER` with the `TokenConfig.getTokenHeaderNae()`.
This extracts the token header name from the settings.

The `TokenUtility` is intended to be used statically.
The `@Configuration` does not operate statically, normally.
Utilize non-static setting configurations to initialize the static variables.

Provide unit tests.
The properties must be tested using separate tests based on on spring boot tests work.
Provide an abstract base test to reduce code repetition in tests that require different properties.
This allows to test and confirm that assigning different header names and cookie names works as intended.