RTR 278-286: Add s to array function names and add more unit tests.
Change the function names that accept different arguments to make this safer for the different languages that might use this utility.
Use the plural `s` to refer to processing an array of headers.

Update the unit tests to handle all of the functions.
Rename the tests to not use the word `set` now that there are plural function names.

The `parseCookieName()` can throw an exception on an empty header string due to the design of `HttpCookie`.
If the header string is empty, then this should not be an error.
Just return no results (`null`) because this is a correct response in this case rather than throwing an exception.