RTR 278-293: Implement support for asArray in RequestDelegate.
Clean up the `HeaderOutputVariables` processing code.

Pass along an Array List of strings to represent the array values.
Conditionally apply spin if requested.

Add `DelegateSpinFailure` as a specific runtime type of exception for the spin operation.
A runtime exception is required because `JsonProcessingException` will not bubble upward in this specific case.